abstract module now have an opacity, giving their cost when called
The opacity of a module is either Opaque (default) or Open.
Essentially, the opacity tells what is the intrinsic cost of an
abstract module when it is called.

If `A` has type `M[f : [c, ...]]` then:
- if A module type is opaque, then calling A.f costs `[A.f : '1]`
- if A module type is open, then calling A.f costs `c`

Opacity is denoted by adding the `open` or `opaque` keyword in a
module complexity restriction, e.g. `M[open f : ...]`.
No explicit opacity means Opaque.